Initialize thread create/stop inside start and stop of capturer/renderer pipeline
Refactored the custom audio device class

Added README for exampleCustomAudioDevice

Create the custom audio device once

Remove dead code

Updated the diagram

Commented FileAndMicAudioDevice class

Use Voice Android SDK 5.2.0

Minor readme update

Minor readme update
