+++ app/build.gradle
@@ -69,7 +69,7 @@ android {
 dependencies {
     testImplementation 'junit:junit:4.12'
 
-    implementation 'com.twilio:voice-android:5.1.1'
+    implementation 'com.twilio:voice-android:5.2.0'
     implementation 'com.android.support:design:28.0.0'
     implementation 'com.android.support:support-media-compat:28.0.0'
     implementation 'com.android.support:animated-vector-drawable:28.0.0'
@@ -78,6 +78,7 @@ dependencies {
     implementation 'com.koushikdutta.ion:ion:2.1.8'
     implementation 'com.google.firebase:firebase-messaging:17.6.0'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
 }
 
 apply plugin: 'com.google.gms.google-services'
