+++ exampleCustomAudioDevice/build.gradle
@@ -45,7 +45,7 @@ android {
 dependencies {
     testImplementation 'junit:junit:4.12'
 
-    implementation 'com.twilio:voice-android:5.1.1'
+    implementation 'com.twilio:voice-android:5.2.0'
     implementation 'com.android.support:design:28.0.0'
     implementation 'com.android.support:support-media-compat:28.0.0'
     implementation 'com.android.support:animated-vector-drawable:28.0.0'
@@ -54,4 +54,5 @@ dependencies {
     implementation 'com.koushikdutta.ion:ion:2.1.8'
     implementation 'com.google.firebase:firebase-messaging:17.6.0'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
 }
