+++ exampleCustomAudioDevice/src/androidTest/java/com/twilio/examplecustomaudiodevice/ExampleInstrumentedTest.java
@@ -1,14 +1,7 @@
 package com.twilio.examplecustomaudiodevice;
 
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
-
-import org.junit.Test;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import org.junit.runner.RunWith;
-
-import static org.junit.Assert.*;
-
 /**
  * Instrumented test, which will execute on an Android device.
  *
@@ -16,11 +9,4 @@ import static org.junit.Assert.*;
  */
 @RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-
-        assertEquals("com.twilio.examplecustomaudiodevice", appContext.getPackageName());
-    }
 }
