+++ exampleCustomAudioDevice/src/test/java/com/twilio/examplecustomaudiodevice/ExampleUnitTest.java
@@ -10,8 +10,5 @@ import static org.junit.Assert.*;
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
 public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
+
 }
